+++ metadata.rb
@@ -14,7 +14,7 @@ issues_url 'https://github.com/socrata-cookbooks/openvpn_okta/issues'
 chef_version '>= 12.1'
 
 depends 'packagecloud', '< 2.0'
-depends 'openvpn', '~> 2.1'
+depends 'openvpn', '~> 3.0'
 
 supports 'ubuntu'
 supports 'redhat'