+++ test/fixtures/cookbooks/test/recipes/default.rb
@@ -2,10 +2,6 @@
 
 apt_update 'default' if platform_family?('debian')
 
-include_recipe 'openvpn_okta'
+directory '/etc/openvpn'
 
-if platform_family?('rhel') && node['platform_version'].to_i >= 7
-  edit_resource! :service, 'openvpn' do
-    service_name 'openvpn@server'
-  end
-end
+include_recipe 'openvpn_okta'